requirements.txt
flask==3.1.0
flask-restx==1.3.0
pydantic==2.9.2
python-dotenv==1.0.1
requests==2.32.3
pyjwt==2.9.0